+++ tests/testthat/test-krig_lst.R
@@ -8,7 +8,7 @@ df <- rbind(df_gx, df_gy)
 
 
 
-context("test-krig_lst.R")
+context("krig.R")
 
 # Keep despite guess_plotdim() is externa. What matters here is not the function
 # but the value it returs. Else, the output of krig and friends will change.
@@ -17,7 +17,7 @@ test_that("plotdimensions are guessed correctly", {
 })
 
 vars <- c("c", "p")
-out_lst <- krig_lst(soil_fake, vars, quiet = TRUE)
+out_lst <- krig(soil_fake, vars, quiet = TRUE)
 
 test_that("outputs object of expected structure at the surface of the list", {
   expect_equal(names(out_lst), vars)
@@ -25,19 +25,19 @@ test_that("outputs object of expected structure at the surface of the list", {
   expect_equal(class(out_lst), classes)
 })
 
-result <- krig_lst(df, var = "m3al", quiet = TRUE)[[1]]
+result <- krig(df, var = "m3al", quiet = TRUE)[[1]]
 
 test_that("fails if var is of length greater than 1", {
   expect_error(
-    krig_lst(df, var = c("m3al", "wrong_name"), quiet = TRUE)[[1]], 
+    krig(df, var = c("m3al", "wrong_name"), quiet = TRUE)[[1]], 
     "isn't in your data"
   )
 })
 
 
 test_that("keeps quiet if asked to", {
-  expect_message(krig_lst(df, var = "m3al"))
-  expect_silent(krig_lst(df, var = "m3al", quiet = TRUE))
+  expect_message(krig(df, var = "m3al"))
+  expect_silent(krig(df, var = "m3al", quiet = TRUE))
 })
 
 test_that("passes regression test", {
@@ -65,10 +65,10 @@ test_that("returns the expected value.", {
 
 test_that("outputs the same with plotdim given directly or via guess_plotdim", {
   expect_equal(
-    krig_lst(
+    krig(
       soil_random, "m3al", quiet = TRUE, plotdim = c(1000, 500),
     )[[1]],
-    krig_lst(
+    krig(
       soil_random, "m3al", quiet = TRUE, plotdim = guess_plotdim(soil_random),
     )[[1]]
   )
@@ -76,34 +76,34 @@ test_that("outputs the same with plotdim given directly or via guess_plotdim", {
 
 test_that("check_GetKrigSoil() fails with wrong input", {
   numeric_input <- as.matrix(df)
-  expect_error(krig_lst(numeric_input, var = "m3al"), "is not TRUE")
+  expect_error(krig(numeric_input, var = "m3al"), "is not TRUE")
 
   rnm <- stats::setNames(df, c("wrong_x", "wrong_gy", "m3al"))
   expect_error(
-    krig_lst(rnm, var = "m3al"),
+    krig(rnm, var = "m3al"),
     "Ensure your data set has these variables:"
   )
   cero_row <- data.frame(gx = numeric(0), gy = numeric(0))
   expect_error(
-    krig_lst(cero_row, var = "m3al"),
+    krig(cero_row, var = "m3al"),
     "Ensure `df.soil` has one or more rows"
   )
   expect_error(
-    krig_lst(df, var = "non-existent-var"),
+    krig(df, var = "non-existent-var"),
     "The variable-name passed to `var` isn't in your data"
   )
-  expect_error(krig_lst(df, var = 888), "is not TRUE")
+  expect_error(krig(df, var = 888), "is not TRUE")
   expect_error(
-    krig_lst(df),
+    krig(df),
     "argument \"var\" is missing"
   )
-  expect_error(krig_lst(df, var = "m3al", gridSize = "3"))
-  expect_error(krig_lst(df, var = "m3al", xSize = "3"))
-  expect_error(krig_lst(df, var = "m3al", ySize = "3"))
+  expect_error(krig(df, var = "m3al", gridSize = "3"))
+  expect_error(krig(df, var = "m3al", xSize = "3"))
+  expect_error(krig(df, var = "m3al", ySize = "3"))
   wrong_type <- 1
-  expect_error(krig_lst(df, var = "m3al", params = wrong_type))
+  expect_error(krig(df, var = "m3al", params = wrong_type))
   bad_not_a_number <- "a"
-  expect_error(krig_lst(df, var = "m3al", breaks = bad_not_a_number))
+  expect_error(krig(df, var = "m3al", breaks = bad_not_a_number))
   bad_not_logical <- "a"
-  expect_error(krig_lst(df, var = "m3al", useKsLine = bad_not_logical))
+  expect_error(krig(df, var = "m3al", useKsLine = bad_not_logical))
 })