+++ tests/testthat/test-summary.krig_lst.R
@@ -1,7 +1,7 @@
-context("test-summary.krig_lst.R")
+context("summary.krig.R")
 
 test_that("outputs correct names", {
-  res <- krig_lst(soil_fake, "c", quiet = TRUE)
+  res <- krig(soil_fake, "c", quiet = TRUE)
   out <- utils::capture.output(summary(res))
   correct_names <- names(res[[1]]) %>% 
     purrr::map_lgl(~any(suppressWarnings(stringr::str_detect(.x, out)))) %>%